Avoid duplicating code in dns_injection_impl and tcp_connect_impl moving it to the base class ooni_test_impl
