+++ src/ooni/dns_injection_impl.hpp
@@ -7,10 +7,8 @@
 
 #include <event2/dns.h>
 #include <measurement_kit/dns.hpp>
-#include "src/ooni/errors.hpp"
 #include "src/ooni/ooni_test_impl.hpp"
 #include "src/ooni/dns_test_impl.hpp"
-#include <sys/stat.h>
 
 namespace mk {
 namespace ooni {
@@ -24,14 +22,7 @@ class DNSInjectionImpl : public DNSTestImpl {
         test_name = "dns_injection";
         test_version = "0.0.1";
 
-        if (input_filepath_ == "") {
-            throw InputFileRequired("An input file is required!");
-        }
-
-        struct stat buffer;
-        if (stat(input_filepath_.c_str(), &buffer) != 0) {
-            throw InputFileDoesNotExist(input_filepath_ + " does not exist");
-        }
+        validate_input_filepath(input_filepath_);
     };
 
     void main(std::string input, Settings options,
